clinker-core: extract Combine dispatch arm to executor/combine_dispatch.rs
Move the PlanNode::Combine arm body — strategy selection (inline hash
build/probe, grace-hash, sort-merge, IEJoin), build/probe orchestration,
the per-driver match/emit loop, and the recoverable output-row error
helper dispatch_combine_output_error — out of executor/dispatch.rs into a
new sibling module executor/combine_dispatch.rs. The dispatcher's Combine
arm is now a single delegating call into combine_dispatch::dispatch_combine.

Pure relocation: no symbol renames, no public-API change. Cross-arm
helpers the moved body still calls (record_error_to_buffer_if_grouped,
tee_emit_to_region_input_buffers, stream_linear_producer_emit) are widened
from file-private to pub(crate) so they resolve across the new module
boundary; they keep real non-test intra-crate callers in dispatch.rs and
the other dispatch arms.

Closes #245
